Break landing page into components
